+++ cargo-dist/src/main.rs
@@ -304,9 +304,9 @@ fn default_desktop_targets() -> Vec<String> {
         "x86_64-unknown-linux-gnu".to_owned(),
         "x86_64-apple-darwin".to_owned(),
         "x86_64-pc-windows-msvc".to_owned(),
+        "aarch64-apple-darwin".to_owned(),
         // cross-compiles not yet supported
         // "aarch64-gnu-unknown-linux".to_owned(),
         // "aarch64-pc-windows-msvc".to_owned(),
-        // "aarch64-apple-darwin".to_owned(),
     ]
 }